LakeCTF2023_quals/web-GeoGuessy/requirements.txt
ngrok
Flask
python-socketio
beautifulsoup4
requests
